src/requirements.txt
langchain
chromadb
openai
python-dotenv
unstructured[local-inference]
wheel
layoutparser[layoutmodels,tesseract]
networkx
cohere